move camera related stuff into it's own package
