+++ chunky/src/java/se/llbit/chunky/renderer/PathTracingRenderer.java
@@ -16,7 +16,7 @@
  */
 package se.llbit.chunky.renderer;
 
-import se.llbit.chunky.renderer.scene.Camera;
+import se.llbit.chunky.renderer.scene.camera.Camera;
 import se.llbit.chunky.renderer.scene.RayTracer;
 import se.llbit.chunky.renderer.scene.Scene;
