+++ chunky/src/java/se/llbit/chunky/renderer/scene/camera/projection/OmniDirectionalStereoProjector.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with Chunky.  If not, see <http://www.gnu.org/licenses/>.
  */
-package se.llbit.chunky.renderer.projection;
+package se.llbit.chunky.renderer.scene.camera.projection;
 
 import org.apache.commons.math3.util.FastMath;
 import se.llbit.math.Constants;
