+++ chunky/src/java/se/llbit/chunky/ui/ChunkMap.java
@@ -35,9 +35,9 @@ import javafx.stage.PopupWindow;
 import se.llbit.chunky.map.MapBuffer;
 import se.llbit.chunky.map.MapView;
 import se.llbit.chunky.map.WorldMapLoader;
-import se.llbit.chunky.renderer.CameraViewListener;
+import se.llbit.chunky.renderer.scene.camera.CameraViewListener;
 import se.llbit.chunky.renderer.ChunkViewListener;
-import se.llbit.chunky.renderer.scene.Camera;
+import se.llbit.chunky.renderer.scene.camera.Camera;
 import se.llbit.chunky.renderer.scene.SceneManager;
 import se.llbit.chunky.ui.controller.ChunkyFxController;
 import se.llbit.chunky.world.Chunk;
