+++ chunky/src/java/se/llbit/chunky/ui/RenderCanvasFx.java
@@ -50,7 +50,7 @@ import javafx.stage.PopupWindow;
 import se.llbit.chunky.PersistentSettings;
 import se.llbit.chunky.renderer.*;
 import se.llbit.chunky.renderer.RenderManager;
-import se.llbit.chunky.renderer.scene.Camera;
+import se.llbit.chunky.renderer.scene.camera.Camera;
 import se.llbit.chunky.renderer.scene.Scene;
 import se.llbit.chunky.ui.controller.ChunkyFxController;
 import se.llbit.math.Vector2;
