+++ chunky/src/java/se/llbit/chunky/ui/render/tabs/CameraTab.java
@@ -37,11 +37,11 @@ import javafx.scene.input.KeyEvent;
 import javafx.stage.FileChooser;
 import se.llbit.chunky.PersistentSettings;
 import se.llbit.chunky.map.MapView;
-import se.llbit.chunky.renderer.ApertureShape;
-import se.llbit.chunky.renderer.CameraViewListener;
-import se.llbit.chunky.renderer.projection.ProjectionMode;
-import se.llbit.chunky.renderer.scene.Camera;
-import se.llbit.chunky.renderer.scene.CameraPreset;
+import se.llbit.chunky.renderer.scene.camera.ApertureShape;
+import se.llbit.chunky.renderer.scene.camera.CameraViewListener;
+import se.llbit.chunky.renderer.scene.camera.projection.ProjectionMode;
+import se.llbit.chunky.renderer.scene.camera.Camera;
+import se.llbit.chunky.renderer.scene.camera.CameraPreset;
 import se.llbit.chunky.renderer.scene.Scene;
 import se.llbit.chunky.ui.DoubleAdjuster;
 import se.llbit.chunky.ui.DoubleTextField;
